Time: 9 ms (40.89%) | Memory: 49.3 MB (7.78%) - LeetSync

+++ 1116-maximum-level-sum-of-a-binary-tree/maximum-level-sum-of-a-binary-tree.java
@@ -21,13 +21,13 @@ class Solution {
         Queue<TreeNode>q =new LinkedList<>();
         q.offer(root);
 
-        int maxsum =Integer.MIN_VALUE;
+        long maxsum =Integer.MIN_VALUE;
         int maxlevel =1;
         int currentlevel =1;
 
         while(!q.isEmpty()){
             int levelsize =q.size();
-            int sum =0;
+            long sum =0;
 
             for(int i =0;i<levelsize;i++){
                 TreeNode node =q.poll();
